+++ lib/src/flutter/context_extensions/media_query_extension.dart
@@ -1,5 +1,3 @@
-import 'dart:ui';
-
 import 'package:flutter/widgets.dart';
 
 import 'package:awesome_extensions/awesome_extensions_flutter.dart';