+++ lib/src/flutter/context_extensions/theme_extension.dart
@@ -4,13 +4,7 @@ extension ThemeExt on BuildContext {
   //  THEMES
 
   /// performs a simple [Theme.of(context).appBarTheme] action and returns given [appBarTheme]
-  AppBarTheme get appBarTheme => Theme.of(this).appBarTheme;
-
-  /// performs a simple [Theme.of(context).backgroundColor] action and returns given [backgroundColor]
-  Color get backgroundColor => Theme.of(this).colorScheme.surface;
-
-  /// performs a simple [Theme.of(context).backgroundColor] action and returns given [backgroundColor]
-  Color get surface => Theme.of(this).colorScheme.surface;
+  AppBarThemeData get appBarTheme => Theme.of(this).appBarTheme;
 
   /// Middle size of the body styles.
   ///
@@ -32,7 +26,8 @@ extension ThemeExt on BuildContext {
   TextStyle? get bodySmall => textTheme.bodySmall;
 
   /// performs a simple [Theme.of(context).bottomAppBarTheme] action and returns given [bottomAppBarTheme]
-  BottomAppBarTheme get bottomAppBarTheme => Theme.of(this).bottomAppBarTheme;
+  BottomAppBarThemeData get bottomAppBarTheme =>
+      Theme.of(this).bottomAppBarTheme;
 
   // COLOR